chore(profilestart): start the profile on profilehomescreen again
To avoid duplicate logic in the HeatingScreen and the PurgeScreen to
start the shot, the logic goes back to profileHomeScreen.
